+++ src/main/java/com/grupo7/peter_parking/model/Parquimetro.java
@@ -6,7 +6,6 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.DBRef;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -23,7 +22,7 @@ public class Parquimetro {
 
     private Long duracaoEmHoras;
 
-    private BigDecimal valorTotal;
+    private double valorTotal;
 
     @DBRef
     private Carro carro;
